Implemented JWT authentication  and authorization

+++ backend/controllers/goalControllers.js
@@ -1,10 +1,10 @@
 const asyncHandler = require("express-async-handler")
 const Goal  = require("../models/goalModel")
-
+const User = require("../models/userModel")
 
 
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find({})
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json( goals )
 })
 
@@ -16,18 +16,31 @@ const postGoals = asyncHandler(async(req, res) => {
         throw new Error("Please add a text field")
     }
     const goal = await Goal.create({
-        text: text
+        text: text,
+        user:req.user.id
     })
     res.status(200).json({goal})
 })
 
 const updateGoals = asyncHandler(async(req, res) => {
     const {id} = req.params;
-    const {text} = req.body
+  
     const goal = await Goal.findById(id);
     if(!goal){
         res.status(400)
         throw new Error("Goal not found")
+    }
+    const user = await User.findById(req.user.id);
+    // Check for user
+    if(!user){
+       res.status(401)
+       throw new Error("User not found")
+    }
+    //  Make sure the logged in user matches the goal uset
+    if (goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
     }
     const updatedGoal = await Goal.findByIdAndUpdate(id,req.body,{
         new:true,
@@ -45,6 +58,18 @@ const deleteGoals = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new Error("Goal not found")
+    }
+    const user = await User.findById(req.user.id);
+    // Check for user
+    if(!user){
+       res.status(401)
+       throw new Error("User not found")
+    }
+    //  Make sure the logged in user matches the goal uset
+    if (goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
     }
     const deletedGoal  = await Goal.findByIdAndDelete(id)
     res.status(200).json({id: req.params.id })
